Start separate module to test for branch outlet backpool error.
